refactor: split bot.go into focused files
Split the monolithic 1291-line bot.go into 5 focused files:

- bot.go (268) — struct, constructor, send helpers, wizard state
- commands.go (390) — slash command handlers
- callbacks.go (213) — callback routing and handlers
- wizard.go (373) — search creation wizard state machine
- history.go (119) — history pagination

Also fixes 3 CodeRabbit review issues:
- Handle CountSearches errors in handleWatch and onShareCopy
  instead of silently ignoring them
- Guard onConfirm against stale/expired wizard sessions

Closes #175
